Fix Minover with bias parameter
When computing the norm of the current weights we forgot to take into
account the d_bias value.
Note that in this case computing the generalized error does not make
sense since then you know in advance that the set is linearly separable.
This probably means that we should extract the computation of a biased
parameter into a separate (wrapper?) class.

Currently only the Rosenblatt implementation references d_bias when
checking the local potentials. It should be possible to extract this to
the Perceptron class so that it is more reusable.